+++ sources/org/telegram/ui/Components/LayoutHelper.java
@@ -7,11 +7,11 @@ import android.widget.LinearLayout;
 import android.widget.RelativeLayout;
 import org.telegram.messenger.AndroidUtilities;
 import org.telegram.messenger.LocaleController;
-
+/* loaded from: classes3.dex */
 public class LayoutHelper {
     private static int getSize(float f) {
         if (f >= 0.0f) {
-            f = (float) AndroidUtilities.dp(f);
+            f = AndroidUtilities.dp(f);
         }
         return (int) f;
     }
@@ -26,21 +26,21 @@ public class LayoutHelper {
     }
 
     public static FrameLayout.LayoutParams createScroll(int i, int i2, int i3) {
-        return new FrameLayout.LayoutParams(getSize((float) i), getSize((float) i2), i3);
+        return new FrameLayout.LayoutParams(getSize(i), getSize(i2), i3);
     }
 
     public static FrameLayout.LayoutParams createFrame(int i, float f, int i2, float f2, float f3, float f4, float f5) {
-        FrameLayout.LayoutParams layoutParams = new FrameLayout.LayoutParams(getSize((float) i), getSize(f), i2);
+        FrameLayout.LayoutParams layoutParams = new FrameLayout.LayoutParams(getSize(i), getSize(f), i2);
         layoutParams.setMargins(AndroidUtilities.dp(f2), AndroidUtilities.dp(f3), AndroidUtilities.dp(f4), AndroidUtilities.dp(f5));
         return layoutParams;
     }
 
     public static FrameLayout.LayoutParams createFrame(int i, int i2, int i3) {
-        return new FrameLayout.LayoutParams(getSize((float) i), getSize((float) i2), i3);
+        return new FrameLayout.LayoutParams(getSize(i), getSize(i2), i3);
     }
 
     public static FrameLayout.LayoutParams createFrame(int i, float f) {
-        return new FrameLayout.LayoutParams(getSize((float) i), getSize(f));
+        return new FrameLayout.LayoutParams(getSize(i), getSize(f));
     }
 
     public static FrameLayout.LayoutParams createFrame(float f, float f2, int i) {
@@ -71,73 +71,73 @@ public class LayoutHelper {
         if (i6 >= 0 && i7 >= 0) {
             layoutParams.addRule(i6, i7);
         }
-        layoutParams.leftMargin = AndroidUtilities.dp((float) i);
-        layoutParams.topMargin = AndroidUtilities.dp((float) i2);
-        layoutParams.rightMargin = AndroidUtilities.dp((float) i3);
-        layoutParams.bottomMargin = AndroidUtilities.dp((float) i4);
+        layoutParams.leftMargin = AndroidUtilities.dp(i);
+        layoutParams.topMargin = AndroidUtilities.dp(i2);
+        layoutParams.rightMargin = AndroidUtilities.dp(i3);
+        layoutParams.bottomMargin = AndroidUtilities.dp(i4);
         return layoutParams;
     }
 
     public static RelativeLayout.LayoutParams createRelative(int i, int i2, int i3, int i4, int i5, int i6, int i7) {
-        return createRelative((float) i, (float) i2, i3, i4, i5, i6, i7, -1, -1);
+        return createRelative(i, i2, i3, i4, i5, i6, i7, -1, -1);
     }
 
     public static RelativeLayout.LayoutParams createRelative(int i, int i2) {
-        return createRelative((float) i, (float) i2, 0, 0, 0, 0, -1, -1, -1);
+        return createRelative(i, i2, 0, 0, 0, 0, -1, -1, -1);
     }
 
     public static RelativeLayout.LayoutParams createRelative(int i, int i2, int i3) {
-        return createRelative((float) i, (float) i2, 0, 0, 0, 0, i3, -1, -1);
+        return createRelative(i, i2, 0, 0, 0, 0, i3, -1, -1);
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, float f, int i3, int i4, int i5, int i6, int i7) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2), f);
-        layoutParams.setMargins(AndroidUtilities.dp((float) i4), AndroidUtilities.dp((float) i5), AndroidUtilities.dp((float) i6), AndroidUtilities.dp((float) i7));
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(i), getSize(i2), f);
+        layoutParams.setMargins(AndroidUtilities.dp(i4), AndroidUtilities.dp(i5), AndroidUtilities.dp(i6), AndroidUtilities.dp(i7));
         layoutParams.gravity = i3;
         return layoutParams;
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, float f, int i3, int i4, int i5, int i6) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2), f);
-        layoutParams.setMargins(AndroidUtilities.dp((float) i3), AndroidUtilities.dp((float) i4), AndroidUtilities.dp((float) i5), AndroidUtilities.dp((float) i6));
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(i), getSize(i2), f);
+        layoutParams.setMargins(AndroidUtilities.dp(i3), AndroidUtilities.dp(i4), AndroidUtilities.dp(i5), AndroidUtilities.dp(i6));
         return layoutParams;
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, int i3, int i4, int i5, int i6, int i7) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2));
-        layoutParams.setMargins(AndroidUtilities.dp((float) i4), AndroidUtilities.dp((float) i5), AndroidUtilities.dp((float) i6), AndroidUtilities.dp((float) i7));
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(i), getSize(i2));
+        layoutParams.setMargins(AndroidUtilities.dp(i4), AndroidUtilities.dp(i5), AndroidUtilities.dp(i6), AndroidUtilities.dp(i7));
         layoutParams.gravity = i3;
         return layoutParams;
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, float f, float f2, float f3, float f4) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2));
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(i), getSize(i2));
         layoutParams.setMargins(AndroidUtilities.dp(f), AndroidUtilities.dp(f2), AndroidUtilities.dp(f3), AndroidUtilities.dp(f4));
         return layoutParams;
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, float f, int i3) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2), f);
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(i), getSize(i2), f);
         layoutParams.gravity = i3;
         return layoutParams;
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, int i3) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2));
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(i), getSize(i2));
         layoutParams.gravity = i3;
         return layoutParams;
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2, float f) {
-        return new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2), f);
+        return new LinearLayout.LayoutParams(getSize(i), getSize(i2), f);
     }
 
     public static LinearLayout.LayoutParams createLinear(int i, int i2) {
-        return new LinearLayout.LayoutParams(getSize((float) i), getSize((float) i2));
+        return new LinearLayout.LayoutParams(getSize(i), getSize(i2));
     }
 
     public static LinearLayout.LayoutParams createLinearRelatively(float f, float f2, int i, float f3, float f4, float f5, float f6) {
-        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(f), getSize(f2), (float) getAbsoluteGravity(i));
+        LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(getSize(f), getSize(f2), getAbsoluteGravity(i));
         layoutParams.leftMargin = AndroidUtilities.dp(LocaleController.isRTL ? f5 : f3);
         layoutParams.topMargin = AndroidUtilities.dp(f4);
         if (!LocaleController.isRTL) {
